pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-plugin-tester"
version = "0.1.0"
description = "Automated security testing tool for detecting over-permissions and data exfiltration in LLM plugins (GPT Actions, MCP Servers, LangChain tools)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Bug Hunter", email = "security@example.com"}
]
keywords = ["llm", "security", "bug-bounty", "plugin", "gpt-actions", "mcp", "langchain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "rich>=13.7.0",
    "loguru>=0.7.2",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "typer>=0.9.0",
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]
proxy = [
    "mitmproxy>=10.1.0",
]
fuzzing = [
    "promptfoo>=0.40.0",
]

[project.scripts]
llm-plugin-tester = "llm_plugin_tester.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/llm-plugin-tester"
Documentation = "https://github.com/yourusername/llm-plugin-tester#readme"
Repository = "https://github.com/yourusername/llm-plugin-tester"
Issues = "https://github.com/yourusername/llm-plugin-tester/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src/llm_plugin_tester --cov-report=term-missing"
